fix:dulication of code fixed
