+++ src/domains/models/Preview.ts
@@ -1,5 +1,5 @@
 import { SpecificationFileNotFound } from '@errors/specification-file';
-import { existsSync,readFileSync } from 'fs';
+import { readFileSync } from 'fs';
 import bundle from '@asyncapi/bundler';
 import { createServer } from 'http';
 import { WebSocketServer } from 'ws';
@@ -8,51 +8,21 @@ import open from 'open';
 import path from 'path';
 import yaml from 'js-yaml';
 import { blueBright, redBright } from 'picocolors';
+import {
+  DEFAULT_PORT,
+  getStudioVersion,
+  isValidFilePath,
+  resolveStudioNextInstance,
+  resolveStudioPath,
+} from '@models/studio-runtime';
+
+export { DEFAULT_PORT } from '@models/studio-runtime';
 
 const sockets: any[] = [];
 const messageQueue: string[] = [];
 const filePathsToWatch: Set<string> = new Set<string>();
 const defaultErrorMessage = 'error occured while bundling files. use --detailedLog or -l flag to get more details.';
 
-export const DEFAULT_PORT = 0;
-
-function isValidFilePath(filePath: string): boolean {
-  return existsSync(filePath);
-}
-
-type NextFactory = (config?: any) => any;
-
-// @asyncapi/studio is an optional, on-demand dependency (installed on first use
-// into the CLI data directory, see studio-installer.ts). Resolving it lazily
-// keeps unrelated commands working when Studio is not installed.
-function resolveStudioPath(): string {
-  try {
-    return path.dirname(require.resolve('@asyncapi/studio/package.json'));
-  } catch {
-    throw new Error(
-      'Preview is not available in this installation. Run the command in an interactive terminal to install Studio on-demand, or pass "--yes" to install automatically.',
-    );
-  }
-}
-
-function getStudioVersion(studioPath?: string): string {
-  try {
-    const pkgPath = studioPath
-      ? path.join(studioPath, 'package.json')
-      : require.resolve('@asyncapi/studio/package.json');
-    return require(pkgPath).version;
-  } catch {
-    return 'unknown';
-  }
-}
-
-// Using require here is necessary for dynamic module resolution
-function resolveStudioNextInstance(studioPath: string): NextFactory {
-  const resolvedNextPath = require.resolve('next', { paths: [studioPath] });
-  const nextModule = require(resolvedNextPath);
-  return nextModule.default ?? nextModule;
-}
- 
 export function startPreview(filePath:string,base:string | undefined,baseDirectory:string | undefined ,xOrigin:boolean | undefined,suppressLogs:boolean|undefined,port: number = DEFAULT_PORT, noBrowser?: boolean, studioPath?: string):void {
   if (filePath && !isValidFilePath(filePath)) {
     throw new SpecificationFileNotFound(filePath);
@@ -61,7 +31,7 @@ export function startPreview(filePath:string,base:string | undefined,baseDirecto
   const resolvedFilePath = path.resolve(filePath);
   const baseDir = path.dirname(resolvedFilePath);
 
-  const resolvedStudioPath = studioPath ?? resolveStudioPath();
+  const resolvedStudioPath = studioPath ?? resolveStudioPath('Preview');
   const nextInstance = resolveStudioNextInstance(resolvedStudioPath);
   const app = nextInstance({
     dev: false,
