+++ src/domains/models/Studio.ts
@@ -1,57 +1,25 @@
-import { existsSync, promises as fPromises } from 'fs';
+import { promises as fPromises } from 'fs';
 import { SpecificationFileNotFound } from '@errors/specification-file';
 import { createServer } from 'http';
 import { WebSocketServer } from 'ws';
 import chokidar from 'chokidar';
 import open from 'open';
-import path from 'path';
 import { blueBright, redBright } from 'picocolors';
+import {
+  DEFAULT_PORT,
+  getStudioVersion,
+  isValidFilePath,
+  resolveStudioNextInstance,
+  resolveStudioPath,
+} from '@models/studio-runtime';
+
+export { DEFAULT_PORT } from '@models/studio-runtime';
 
 const { readFile, writeFile } = fPromises;
 
 const sockets: any[] = [];
 const messageQueue: string[] = [];
 
-export const DEFAULT_PORT = 0;
-
-function isValidFilePath(filePath: string): boolean {
-  return existsSync(filePath);
-}
-
-type NextFactory = (config?: any) => any;
-
-// @asyncapi/studio is an optional, on-demand dependency (installed on first use
-// into the CLI data directory, see studio-installer.ts). Resolving it lazily
-// means commands which only reference this module (like `new file` without
-// `--studio`) keep working even when Studio is not installed.
-function resolveStudioPath(): string {
-  try {
-    return path.dirname(require.resolve('@asyncapi/studio/package.json'));
-  } catch {
-    throw new Error(
-      'Studio is not available in this installation. Run the command in an interactive terminal to install it on-demand, or pass "--yes" to install automatically.',
-    );
-  }
-}
-
-function getStudioVersion(studioPath?: string): string {
-  try {
-    const pkgPath = studioPath
-      ? path.join(studioPath, 'package.json')
-      : require.resolve('@asyncapi/studio/package.json');
-    return require(pkgPath).version;
-  } catch {
-    return 'unknown';
-  }
-}
-
-// Using require here is necessary for dynamic module resolution
-function resolveStudioNextInstance(studioPath: string): NextFactory {
-  const resolvedNextPath = require.resolve('next', { paths: [studioPath] });
-  const nextModule = require(resolvedNextPath);
-  return nextModule.default ?? nextModule;
-}
- 
 export function start(filePath: string, port: number = DEFAULT_PORT, noBrowser?:boolean, studioPath?: string): void {
   if (filePath && !isValidFilePath(filePath)) {
     throw new SpecificationFileNotFound(filePath);
@@ -59,7 +27,7 @@ export function start(filePath: string, port: number = DEFAULT_PORT, noBrowser?:
 
   // Locate @asyncapi/studio package (resolved on-demand by the command, or
   // fall back to node_modules resolution for bundled installs).
-  const resolvedStudioPath = studioPath ?? resolveStudioPath();
+  const resolvedStudioPath = studioPath ?? resolveStudioPath('Studio');
   const nextInstance = resolveStudioNextInstance(resolvedStudioPath);
   const app = nextInstance({
     dev: false,
